# Transition complex objects to "too complex" shape