--- a/ext/objspace/objspace_dump.c
+++ b/ext/objspace/objspace_dump.c
@@ -15,6 +15,7 @@
 #include "gc.h"
 #include "internal.h"
 #include "internal/array.h"
+#include "internal/class.h"
 #include "internal/hash.h"
 #include "internal/string.h"
 #include "internal/sanitizers.h"
@@ -498,6 +499,9 @@ dump_object(VALUE obj, struct dump_config *dc)
         break;
 
       case T_CLASS:
+        dump_append(dc, ", \"variation_count\":");
+        dump_append_d(dc, RCLASS_EXT(obj)->variation_count);
+
       case T_MODULE:
         if (rb_class_get_superclass(obj)) {
             dump_append(dc, ", \"superclass\":");