… test/ruby/test_shapes.rb
… variable.c
… yjit/src/codegen.rs