@maximecb 2022-12-14
I ran our headline benchmarks locally:

```
$ ./run_benchmarks.rb --chruby "master::ruby-yjit --yjit" --chruby "complex-shapes::ruby-complex-shapes --yjit" activerecord hexapdf liquid-render mail railsbench ruby-lsp

master: ruby 3.2.0dev (2022-12-14T21:17:46Z master 693c01d509) +YJIT [x86_64-darwin22]
complex-shapes: ruby 3.2.0dev (2022-12-14T19:39:45Z fallback-to-hash 95af8be94f) +YJIT [x86_64-darwin22]

-------------  -----------  ----------  -------------------  ----------  ---------------------  ----------------------
bench          master (ms)  stddev (%)  complex-shapes (ms)  stddev (%)  master/complex-shapes  complex-shapes 1st itr
activerecord   73.4         0.7         72.9                 0.9         1.01                   0.97                  
hexapdf        1635.2       4.1         1652.3               4.2         0.99                   1.04                  
liquid-render  88.0         1.4         88.5                 1.5         0.99                   0.99                  
mail           109.7        0.8         111.8                1.0         0.98                   0.99                  
railsbench     1414.4       1.8         1417.1               2.9         1.00                   1.08                  
ruby-lsp       71.0         53.2        74.6                 49.9        0.95                   0.78                  
-------------  -----------  ----------  -------------------  ----------  ---------------------  ----------------------
Legend:
- master/complex-shapes: ratio of master/complex-shapes time. Higher is better for complex-shapes. Above 1 represents a speedup.
- complex-shapes 1st itr: ratio of master/complex-shapes time for the first benchmarking iteration.
```

There's some noise because I'm just running this on my macbook, and ruby-lsp is especially prone to noise because it uses a server process AFAIK, but the numbers seem generally on par with what they were before 👍