@maximecb 2022-12-14
Avg number of instructions in YJIT also remains unchanged for railsbench 👍