@k0kubun 2022-12-14
> I'm just running this on my macbook

I'll share another data point on a Linux desktop with CPU frequency stabilized :)

```
before: ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
last_commit=Remove dead code in get_next_shape_internal
after: ruby 3.2.0dev (2022-12-14T21:39:52Z fallback-to-hash 4c486ff1d1) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
activerecord   63.7         0.5         64.1        0.5         0.99          0.99
hexapdf        1205.9       1.1         1189.7      1.8         1.01          1.00
liquid-render  65.6         1.3         66.6        1.4         0.99          0.99
mail           77.1         0.5         77.4        0.4         1.00          1.00
psych-load     980.7        0.2         1019.8      0.1         0.96          0.96
railsbench     1072.1       1.3         1065.0      1.7         1.01          1.00
ruby-lsp       39.1         26.0        39.4        29.0        0.99          1.01
-------------  -----------  ----------  ----------  ----------  ------------  -------------
```

<details>

```
$ ./run_benchmarks.rb activerecord hexapdf liquid-render mail psych-load railsbench ruby-lsp --chruby "before::before --yjit" --chruby "after::after --yjit"
Running benchmark "activerecord" (1/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/activerecord/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
-- create_table(:posts, {:force=>true})
   -> 0.0022s
itr #1: 80ms
itr #2: 63ms
itr #3: 63ms
itr #4: 63ms
itr #5: 63ms
itr #6: 63ms
itr #7: 63ms
itr #8: 63ms
itr #9: 63ms
itr #10: 63ms
itr #11: 63ms
itr #12: 63ms
itr #13: 63ms
itr #14: 63ms
itr #15: 63ms
itr #16: 63ms
itr #17: 63ms
itr #18: 63ms
itr #19: 63ms
itr #20: 63ms
itr #21: 63ms
itr #22: 63ms
itr #23: 63ms
itr #24: 63ms
itr #25: 63ms
itr #26: 63ms
itr #27: 63ms
itr #28: 63ms
itr #29: 63ms
itr #30: 63ms
itr #31: 63ms
itr #32: 64ms
itr #33: 64ms
itr #34: 64ms
itr #35: 63ms
itr #36: 63ms
itr #37: 63ms
itr #38: 63ms
itr #39: 63ms
itr #40: 63ms
itr #41: 63ms
itr #42: 63ms
itr #43: 63ms
itr #44: 63ms
itr #45: 63ms
itr #46: 63ms
itr #47: 63ms
itr #48: 63ms
itr #49: 63ms
itr #50: 63ms
itr #51: 63ms
itr #52: 63ms
itr #53: 63ms
itr #54: 63ms
itr #55: 63ms
itr #56: 63ms
itr #57: 63ms
itr #58: 63ms
itr #59: 63ms
itr #60: 63ms
itr #61: 63ms
itr #62: 63ms
itr #63: 63ms
itr #64: 63ms
itr #65: 63ms
itr #66: 63ms
itr #67: 63ms
itr #68: 65ms
itr #69: 64ms
itr #70: 63ms
itr #71: 63ms
itr #72: 63ms
itr #73: 63ms
itr #74: 63ms
itr #75: 63ms
itr #76: 63ms
itr #77: 63ms
itr #78: 63ms
itr #79: 63ms
itr #80: 63ms
itr #81: 63ms
itr #82: 63ms
itr #83: 63ms
itr #84: 63ms
itr #85: 63ms
itr #86: 63ms
itr #87: 63ms
itr #88: 63ms
itr #89: 63ms
itr #90: 63ms
itr #91: 63ms
itr #92: 63ms
itr #93: 63ms
itr #94: 63ms
itr #95: 63ms
itr #96: 63ms
itr #97: 63ms
itr #98: 63ms
itr #99: 63ms
itr #100: 63ms
itr #101: 63ms
itr #102: 63ms
itr #103: 65ms
itr #104: 64ms
itr #105: 63ms
itr #106: 63ms
itr #107: 63ms
itr #108: 63ms
itr #109: 63ms
itr #110: 63ms
itr #111: 63ms
itr #112: 63ms
itr #113: 63ms
itr #114: 63ms
itr #115: 63ms
itr #116: 63ms
itr #117: 63ms
itr #118: 63ms
itr #119: 63ms
itr #120: 63ms
itr #121: 63ms
itr #122: 63ms
itr #123: 63ms
itr #124: 63ms
itr #125: 63ms
itr #126: 63ms
itr #127: 63ms
itr #128: 63ms
itr #129: 63ms
itr #130: 63ms
itr #131: 63ms
itr #132: 63ms
itr #133: 63ms
itr #134: 63ms
itr #135: 63ms
itr #136: 63ms
itr #137: 63ms
itr #138: 64ms
itr #139: 64ms
itr #140: 63ms
itr #141: 63ms
itr #142: 63ms
itr #143: 63ms
itr #144: 63ms
itr #145: 63ms
itr #146: 63ms
itr #147: 63ms
itr #148: 63ms
itr #149: 63ms
itr #150: 63ms
itr #151: 63ms
itr #152: 63ms
itr #153: 63ms
itr #154: 63ms
itr #155: 63ms
itr #156: 63ms
itr #157: 63ms
Average of last 142, non-warmup iters: 63ms
Running benchmark "hexapdf" (2/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/hexapdf/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
itr #1: 1450ms
itr #2: 1276ms
itr #3: 1183ms
itr #4: 1214ms
itr #5: 1230ms
itr #6: 1228ms
itr #7: 1199ms
itr #8: 1207ms
itr #9: 1232ms
itr #10: 1202ms
itr #11: 1234ms
itr #12: 1214ms
itr #13: 1198ms
itr #14: 1206ms
itr #15: 1205ms
itr #16: 1211ms
itr #17: 1222ms
itr #18: 1193ms
itr #19: 1221ms
itr #20: 1202ms
itr #21: 1204ms
itr #22: 1217ms
itr #23: 1173ms
itr #24: 1204ms
itr #25: 1207ms
Average of last 10, non-warmup iters: 1205ms
Running benchmark "liquid-render" (3/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/liquid-render/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
itr #1: 87ms
itr #2: 66ms
itr #3: 65ms
itr #4: 65ms
itr #5: 65ms
itr #6: 65ms
itr #7: 65ms
itr #8: 65ms
itr #9: 65ms
itr #10: 65ms
itr #11: 67ms
itr #12: 65ms
itr #13: 65ms
itr #14: 65ms
itr #15: 65ms
itr #16: 65ms
itr #17: 65ms
itr #18: 65ms
itr #19: 65ms
itr #20: 65ms
itr #21: 67ms
itr #22: 65ms
itr #23: 68ms
itr #24: 67ms
itr #25: 66ms
itr #26: 66ms
itr #27: 66ms
itr #28: 66ms
itr #29: 66ms
itr #30: 65ms
itr #31: 65ms
itr #32: 65ms
itr #33: 65ms
itr #34: 65ms
itr #35: 65ms
itr #36: 65ms
itr #37: 65ms
itr #38: 63ms
itr #39: 65ms
itr #40: 65ms
itr #41: 65ms
itr #42: 65ms
itr #43: 65ms
itr #44: 65ms
itr #45: 65ms
itr #46: 63ms
itr #47: 65ms
itr #48: 65ms
itr #49: 65ms
itr #50: 65ms
itr #51: 65ms
itr #52: 65ms
itr #53: 65ms
itr #54: 63ms
itr #55: 65ms
itr #56: 65ms
itr #57: 67ms
itr #58: 67ms
itr #59: 66ms
itr #60: 66ms
itr #61: 65ms
itr #62: 63ms
itr #63: 65ms
itr #64: 65ms
itr #65: 65ms
itr #66: 65ms
itr #67: 65ms
itr #68: 65ms
itr #69: 65ms
itr #70: 63ms
itr #71: 65ms
itr #72: 65ms
itr #73: 65ms
itr #74: 65ms
itr #75: 65ms
itr #76: 65ms
itr #77: 65ms
itr #78: 63ms
itr #79: 65ms
itr #80: 65ms
itr #81: 65ms
itr #82: 65ms
itr #83: 65ms
itr #84: 65ms
itr #85: 65ms
itr #86: 65ms
itr #87: 63ms
itr #88: 65ms
itr #89: 65ms
itr #90: 65ms
itr #91: 65ms
itr #92: 68ms
itr #93: 67ms
itr #94: 66ms
itr #95: 63ms
itr #96: 65ms
itr #97: 65ms
itr #98: 65ms
itr #99: 65ms
itr #100: 65ms
itr #101: 65ms
itr #102: 65ms
itr #103: 63ms
itr #104: 65ms
itr #105: 65ms
itr #106: 65ms
itr #107: 65ms
itr #108: 65ms
itr #109: 65ms
itr #110: 65ms
itr #111: 63ms
itr #112: 65ms
itr #113: 65ms
itr #114: 65ms
itr #115: 65ms
itr #116: 65ms
itr #117: 65ms
itr #118: 65ms
itr #119: 63ms
itr #120: 65ms
itr #121: 65ms
itr #122: 65ms
itr #123: 65ms
itr #124: 65ms
itr #125: 65ms
itr #126: 68ms
itr #127: 67ms
itr #128: 64ms
itr #129: 66ms
itr #130: 66ms
itr #131: 66ms
itr #132: 65ms
itr #133: 65ms
itr #134: 65ms
itr #135: 65ms
itr #136: 63ms
itr #137: 65ms
itr #138: 65ms
itr #139: 65ms
itr #140: 65ms
itr #141: 65ms
itr #142: 65ms
itr #143: 65ms
itr #144: 65ms
itr #145: 63ms
itr #146: 65ms
itr #147: 65ms
itr #148: 65ms
itr #149: 65ms
itr #150: 65ms
itr #151: 65ms
itr #152: 65ms
itr #153: 63ms
Average of last 138, non-warmup iters: 65ms
Running benchmark "mail" (4/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/mail/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
itr #1: 208ms
itr #2: 77ms
itr #3: 77ms
itr #4: 77ms
itr #5: 77ms
itr #6: 77ms
itr #7: 77ms
itr #8: 77ms
itr #9: 77ms
itr #10: 77ms
itr #11: 77ms
itr #12: 77ms
itr #13: 77ms
itr #14: 77ms
itr #15: 77ms
itr #16: 77ms
itr #17: 77ms
itr #18: 77ms
itr #19: 77ms
itr #20: 77ms
itr #21: 77ms
itr #22: 77ms
itr #23: 77ms
itr #24: 77ms
itr #25: 77ms
itr #26: 77ms
itr #27: 77ms
itr #28: 79ms
itr #29: 78ms
itr #30: 77ms
itr #31: 76ms
itr #32: 76ms
itr #33: 76ms
itr #34: 76ms
itr #35: 76ms
itr #36: 76ms
itr #37: 76ms
itr #38: 76ms
itr #39: 76ms
itr #40: 76ms
itr #41: 77ms
itr #42: 76ms
itr #43: 76ms
itr #44: 76ms
itr #45: 76ms
itr #46: 76ms
itr #47: 76ms
itr #48: 76ms
itr #49: 76ms
itr #50: 76ms
itr #51: 77ms
itr #52: 77ms
itr #53: 76ms
itr #54: 76ms
itr #55: 77ms
itr #56: 77ms
itr #57: 77ms
itr #58: 78ms
itr #59: 77ms
itr #60: 76ms
itr #61: 76ms
itr #62: 76ms
itr #63: 76ms
itr #64: 76ms
itr #65: 76ms
itr #66: 77ms
itr #67: 76ms
itr #68: 76ms
itr #69: 76ms
itr #70: 77ms
itr #71: 76ms
itr #72: 76ms
itr #73: 76ms
itr #74: 76ms
itr #75: 76ms
itr #76: 76ms
itr #77: 76ms
itr #78: 76ms
itr #79: 76ms
itr #80: 77ms
itr #81: 77ms
itr #82: 76ms
itr #83: 76ms
itr #84: 77ms
itr #85: 76ms
itr #86: 77ms
itr #87: 77ms
itr #88: 77ms
itr #89: 77ms
itr #90: 76ms
itr #91: 77ms
itr #92: 76ms
itr #93: 77ms
itr #94: 76ms
itr #95: 76ms
itr #96: 77ms
itr #97: 76ms
itr #98: 77ms
itr #99: 76ms
itr #100: 77ms
itr #101: 77ms
itr #102: 76ms
itr #103: 76ms
itr #104: 76ms
itr #105: 77ms
itr #106: 77ms
itr #107: 76ms
itr #108: 76ms
itr #109: 76ms
itr #110: 76ms
itr #111: 77ms
itr #112: 76ms
itr #113: 76ms
itr #114: 76ms
itr #115: 76ms
itr #116: 77ms
itr #117: 77ms
itr #118: 77ms
itr #119: 77ms
itr #120: 77ms
itr #121: 77ms
itr #122: 77ms
itr #123: 76ms
itr #124: 77ms
itr #125: 77ms
itr #126: 77ms
itr #127: 77ms
itr #128: 76ms
itr #129: 76ms
Average of last 114, non-warmup iters: 77ms
Running benchmark "psych-load" (5/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/psych-load/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
itr #1: 988ms
itr #2: 978ms
itr #3: 977ms
itr #4: 979ms
itr #5: 977ms
itr #6: 981ms
itr #7: 978ms
itr #8: 982ms
itr #9: 980ms
itr #10: 979ms
itr #11: 981ms
itr #12: 979ms
itr #13: 981ms
itr #14: 978ms
itr #15: 981ms
itr #16: 977ms
itr #17: 978ms
itr #18: 982ms
itr #19: 978ms
itr #20: 982ms
itr #21: 980ms
itr #22: 982ms
itr #23: 980ms
itr #24: 982ms
itr #25: 980ms
Average of last 10, non-warmup iters: 980ms
Running benchmark "railsbench" (6/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/railsbench/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Command: bin/rails db:migrate db:seed
Using 100 posts in the database
itr #1: 1209ms
itr #2: 1067ms
itr #3: 1060ms
itr #4: 1066ms
itr #5: 1057ms
itr #6: 1067ms
itr #7: 1056ms
itr #8: 1069ms
itr #9: 1092ms
itr #10: 1066ms
itr #11: 1056ms
itr #12: 1068ms
itr #13: 1087ms
itr #14: 1066ms
itr #15: 1059ms
itr #16: 1098ms
itr #17: 1059ms
itr #18: 1065ms
itr #19: 1090ms
itr #20: 1056ms
itr #21: 1068ms
itr #22: 1057ms
itr #23: 1067ms
itr #24: 1090ms
itr #25: 1067ms
Average of last 10, non-warmup iters: 1072ms
Running benchmark "ruby-lsp" (7/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby --yjit -I ./harness benchmarks/ruby-lsp/benchmark.rb
ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
itr #1: 991ms
itr #2: 104ms
itr #3: 97ms
itr #4: 46ms
itr #5: 44ms
itr #6: 45ms
itr #7: 39ms
itr #8: 44ms
itr #9: 37ms
itr #10: 50ms
itr #11: 38ms
itr #12: 36ms
itr #13: 36ms
itr #14: 35ms
itr #15: 52ms
itr #16: 38ms
itr #17: 36ms
itr #18: 36ms
itr #19: 73ms
itr #20: 36ms
itr #21: 35ms
itr #22: 36ms
itr #23: 36ms
itr #24: 35ms
itr #25: 36ms
itr #26: 35ms
itr #27: 36ms
itr #28: 35ms
itr #29: 36ms
itr #30: 92ms
itr #31: 37ms
itr #32: 36ms
itr #33: 36ms
itr #34: 36ms
itr #35: 73ms
itr #36: 36ms
itr #37: 35ms
itr #38: 36ms
itr #39: 35ms
itr #40: 35ms
itr #41: 36ms
itr #42: 35ms
itr #43: 36ms
itr #44: 35ms
itr #45: 35ms
itr #46: 36ms
itr #47: 35ms
itr #48: 35ms
itr #49: 72ms
itr #50: 37ms
itr #51: 36ms
itr #52: 39ms
itr #53: 41ms
itr #54: 39ms
itr #55: 37ms
itr #56: 35ms
itr #57: 37ms
itr #58: 35ms
itr #59: 36ms
itr #60: 36ms
itr #61: 35ms
itr #62: 35ms
itr #63: 36ms
itr #64: 73ms
itr #65: 39ms
itr #66: 36ms
itr #67: 35ms
itr #68: 36ms
itr #69: 35ms
itr #70: 35ms
itr #71: 35ms
itr #72: 36ms
itr #73: 35ms
itr #74: 36ms
itr #75: 35ms
itr #76: 35ms
itr #77: 36ms
itr #78: 36ms
itr #79: 36ms
itr #80: 44ms
itr #81: 64ms
itr #82: 36ms
itr #83: 34ms
itr #84: 36ms
itr #85: 35ms
itr #86: 35ms
itr #87: 35ms
itr #88: 36ms
itr #89: 35ms
itr #90: 36ms
itr #91: 35ms
itr #92: 35ms
itr #93: 36ms
itr #94: 35ms
itr #95: 35ms
itr #96: 35ms
itr #97: 76ms
itr #98: 37ms
itr #99: 35ms
itr #100: 36ms
itr #101: 35ms
itr #102: 35ms
itr #103: 35ms
itr #104: 36ms
itr #105: 35ms
itr #106: 36ms
itr #107: 36ms
itr #108: 36ms
itr #109: 35ms
itr #110: 36ms
itr #111: 82ms
itr #112: 39ms
itr #113: 36ms
itr #114: 34ms
itr #115: 36ms
itr #116: 35ms
itr #117: 35ms
itr #118: 35ms
itr #119: 36ms
itr #120: 35ms
itr #121: 36ms
itr #122: 35ms
itr #123: 35ms
itr #124: 36ms
itr #125: 36ms
itr #126: 36ms
itr #127: 35ms
itr #128: 73ms
itr #129: 36ms
itr #130: 34ms
itr #131: 36ms
itr #132: 35ms
itr #133: 35ms
itr #134: 35ms
itr #135: 36ms
itr #136: 35ms
itr #137: 36ms
itr #138: 35ms
itr #139: 35ms
itr #140: 36ms
itr #141: 36ms
itr #142: 35ms
itr #143: 35ms
itr #144: 77ms
itr #145: 37ms
itr #146: 35ms
itr #147: 36ms
itr #148: 35ms
itr #149: 35ms
itr #150: 35ms
itr #151: 36ms
itr #152: 35ms
itr #153: 35ms
itr #154: 36ms
itr #155: 35ms
itr #156: 35ms
itr #157: 36ms
itr #158: 74ms
itr #159: 38ms
itr #160: 36ms
itr #161: 35ms
itr #162: 36ms
itr #163: 35ms
itr #164: 36ms
itr #165: 35ms
itr #166: 36ms
itr #167: 35ms
itr #168: 36ms
itr #169: 43ms
itr #170: 39ms
itr #171: 38ms
itr #172: 36ms
itr #173: 36ms
itr #174: 49ms
itr #175: 63ms
itr #176: 36ms
itr #177: 35ms
itr #178: 36ms
itr #179: 35ms
itr #180: 35ms
itr #181: 35ms
itr #182: 36ms
itr #183: 35ms
itr #184: 36ms
itr #185: 35ms
itr #186: 35ms
itr #187: 36ms
itr #188: 36ms
itr #189: 36ms
itr #190: 36ms
itr #191: 76ms
itr #192: 36ms
itr #193: 35ms
itr #194: 36ms
itr #195: 35ms
itr #196: 35ms
itr #197: 35ms
itr #198: 36ms
itr #199: 35ms
itr #200: 35ms
itr #201: 35ms
itr #202: 35ms
itr #203: 35ms
itr #204: 36ms
itr #205: 35ms
itr #206: 36ms
itr #207: 77ms
itr #208: 37ms
itr #209: 35ms
itr #210: 36ms
itr #211: 35ms
itr #212: 35ms
itr #213: 35ms
itr #214: 35ms
itr #215: 35ms
itr #216: 35ms
itr #217: 35ms
itr #218: 35ms
itr #219: 36ms
itr #220: 36ms
itr #221: 35ms
itr #222: 74ms
itr #223: 36ms
itr #224: 34ms
itr #225: 36ms
itr #226: 37ms
itr #227: 42ms
itr #228: 39ms
Average of last 213, non-warmup iters: 39ms
Running benchmark "activerecord" (1/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/after/bin/ruby --yjit -I ./harness benchmarks/activerecord/benchmark.rb
ruby 3.2.0dev (2022-12-14T21:39:52Z fallback-to-hash 4c486ff1d1) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
-- create_table(:posts, {:force=>true})
   -> 0.0022s
itr #1: 81ms
itr #2: 64ms
itr #3: 64ms
itr #4: 64ms
itr #5: 64ms
itr #6: 65ms
itr #7: 64ms
itr #8: 64ms
itr #9: 64ms
itr #10: 63ms
itr #11: 64ms
itr #12: 64ms
itr #13: 64ms
itr #14: 63ms
itr #15: 63ms
itr #16: 64ms
itr #17: 64ms
itr #18: 64ms
itr #19: 63ms
itr #20: 63ms
itr #21: 64ms
itr #22: 64ms
itr #23: 64ms
itr #24: 64ms
itr #25: 64ms
itr #26: 64ms
itr #27: 64ms
itr #28: 63ms
itr #29: 63ms
itr #30: 64ms
itr #31: 64ms
itr #32: 64ms
itr #33: 64ms
itr #34: 64ms
itr #35: 64ms
itr #36: 64ms
itr #37: 64ms
itr #38: 64ms
itr #39: 63ms
itr #40: 64ms
itr #41: 65ms
itr #42: 64ms
itr #43: 64ms
itr #44: 64ms
itr #45: 64ms
itr #46: 64ms
itr #47: 64ms
itr #48: 64ms
itr #49: 63ms
itr #50: 64ms
itr #51: 64ms
itr #52: 64ms
itr #53: 64ms
itr #54: 64ms
itr #55: 64ms
itr #56: 64ms
itr #57: 64ms
itr #58: 64ms
itr #59: 64ms
itr #60: 64ms
itr #61: 63ms
itr #62: 63ms
itr #63: 64ms
itr #64: 64ms
itr #65: 63ms
itr #66: 63ms
itr #67: 63ms
itr #68: 64ms
itr #69: 64ms
itr #70: 64ms
itr #71: 64ms
itr #72: 64ms
itr #73: 64ms
itr #74: 63ms
itr #75: 64ms
itr #76: 65ms
itr #77: 64ms
itr #78: 64ms
itr #79: 64ms
itr #80: 63ms
itr #81: 63ms
itr #82: 63ms
itr #83: 63ms
itr #84: 63ms
itr #85: 63ms
itr #86: 63ms
itr #87: 63ms
itr #88: 64ms
itr #89: 64ms
itr #90: 63ms
itr #91: 63ms
itr #92: 64ms
itr #93: 64ms
itr #94: 64ms
itr #95: 64ms
itr #96: 64ms
itr #97: 64ms
itr #98: 63ms
itr #99: 63ms
itr #100: 64ms
itr #101: 64ms
itr #102: 63ms
itr #103: 64ms
itr #104: 64ms
itr #105: 64ms
itr #106: 63ms
itr #107: 63ms
itr #108: 63ms
itr #109: 64ms
itr #110: 64ms
itr #111: 65ms
itr #112: 64ms
itr #113: 64ms
itr #114: 64ms
itr #115: 64ms
itr #116: 63ms
itr #117: 64ms
itr #118: 64ms
itr #119: 63ms
itr #120: 63ms
itr #121: 63ms
itr #122: 64ms
itr #123: 64ms
itr #124: 64ms
itr #125: 64ms
itr #126: 64ms
itr #127: 64ms
itr #128: 64ms
itr #129: 64ms
itr #130: 64ms
itr #131: 64ms
itr #132: 64ms
itr #133: 63ms
itr #134: 63ms
itr #135: 63ms
itr #136: 63ms
itr #137: 63ms
itr #138: 64ms
itr #139: 64ms
itr #140: 64ms
itr #141: 64ms
itr #142: 64ms
itr #143: 63ms
itr #144: 63ms
itr #145: 63ms
itr #146: 65ms
itr #147: 64ms
itr #148: 64ms
itr #149: 63ms
itr #150: 63ms
itr #151: 63ms
itr #152: 63ms
itr #153: 63ms
itr #154: 64ms
itr #155: 63ms
itr #156: 64ms
Average of last 141, non-warmup iters: 64ms
Running benchmark "hexapdf" (2/7)
setarch x86_64 -R taskset -c 7 /opt/rubies/after/bin/ruby --yjit -I ./harness benchmarks/hexapdf/benchmark.rb
ruby 3.2.0dev (2022-12-14T21:39:52Z fallback-to-hash 4c486ff1d1) +YJIT [x86_64-linux]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
itr #1: 1447ms
itr #2: 1297ms
itr #3: 1164ms
itr #4: 1251ms
itr #5: 1180ms
itr #6: 1185ms
itr #7: 1211ms
itr #8: 1171ms
itr #9: 1198ms
itr #10: 1181ms
itr #11: 1165ms
itr #12: 1211ms
itr #13: 1179ms
itr #14: 1176ms
itr #15: 1200ms
itr #16: 1213ms
itr #17: 1180ms
itr #18: 1170ms
itr #19: 1173ms
…